+++ lirical-exomiser-db-adapter/src/main/java/org/monarchinitiative/lirical/exomiser_db_adapter/model/pathogenicity/MutationTasterScore.java
@@ -17,11 +17,12 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package org.monarchinitiative.lirical.exomiser_db_adapter.model.pathogenicity;
 
 /**
  *
- * @author Jules Jacobsen
+ * @author Jules Jacobsen <jules.jacobsen@sanger.ac.uk>
  */
 public class MutationTasterScore extends BasePathogenicityScore {