+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/LiricalDataResolver.java
@@ -10,6 +10,21 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * A convenience class for accessing resources from LIRICAL data directory.
+ * <p>
+ * The resource files include:
+ * <ul>
+ *     <li>HPO in JSON format (<em>hp.json</em>)</li>
+ *     <li>HPO annotations (<em>phenotype.hpoa</em>)</li>
+ *     <li>HGNC complete set - a table with gene identifiers, symbols, and other metadata (<em>hgnc_complete_set.txt</em>)</li>
+ *     <li>MIM to medgene mapping file (<em>mim2gene_medgen</em>)</li>
+ *     <li>Orpha to gene mapping file (<em>en_product6.xml</em>). Optional</li>
+ *     <li>Jannovar caches (<em>hg(19|38)_(ucsc|refseq).ser</em>)</li>
+ * </ul>
+ * <p>
+ * The resolver checks if all non-optional files are present and will raise an exception otherwise.
+ */
 public class LiricalDataResolver {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(LiricalDataResolver.class);
@@ -27,8 +42,12 @@ public class LiricalDataResolver {
 
     private void checkV1Resources() throws LiricalDataException {
         boolean error = false;
-        List<Path> requiredFiles = List.of(hpoJson(), hgncCompleteSet(), mim2geneMedgen(), phenotypeAnnotations(),
-                hg19RefseqTxDatabase(), hg19UcscTxDatabase(), hg38RefseqTxDatabase(), hg38UcscTxDatabase());
+        List<Path> requiredFiles = List.of(
+                hpoJson(), hgncCompleteSet(), mim2geneMedgen(), phenotypeAnnotations(),
+                hg19RefseqTxDatabase(), hg19RefseqCuratedTxDatabase(), hg19EnsemblTxDatabase(), hg19UcscTxDatabase(),
+                hg38RefseqTxDatabase(), hg38RefseqCuratedTxDatabase(), hg38EnsemblTxDatabase(), hg38UcscTxDatabase()
+        );
+        // Note: we do not require `orpha2gene` in all analyses, hence it is not required!
         for (Path file : requiredFiles) {
             if (!Files.isRegularFile(file)) {
                 LOGGER.error("Missing required file `{}` in `{}`.", file.toFile().getName(), dataDirectory.toAbsolutePath());
@@ -46,6 +65,10 @@ public class LiricalDataResolver {
 
     public Path hpoJson() {
         return dataDirectory.resolve("hp.json");
+    }
+
+    public Path orpha2gene() {
+        return dataDirectory.resolve("en_product6.xml");
     }
 
     public Path hgncCompleteSet() {
@@ -68,23 +91,43 @@ public class LiricalDataResolver {
         return dataDirectory.resolve("hg19_refseq.ser");
     }
 
-    public Path hg38RefseqTxDatabase() {
-        return dataDirectory.resolve("hg38_refseq.ser");
+    public Path hg19RefseqCuratedTxDatabase() {
+        return dataDirectory.resolve("hg19_refseq_curated.ser");
+    }
+
+    public Path hg19EnsemblTxDatabase() {
+        return dataDirectory.resolve("hg19_ensembl.ser");
     }
 
     public Path hg38UcscTxDatabase() {
         return dataDirectory.resolve("hg38_ucsc.ser");
     }
 
+    public Path hg38RefseqTxDatabase() {
+        return dataDirectory.resolve("hg38_refseq.ser");
+    }
+
+    public Path hg38RefseqCuratedTxDatabase() {
+        return dataDirectory.resolve("hg38_refseq_curated.ser");
+    }
+
+    public Path hg38EnsemblTxDatabase() {
+        return dataDirectory.resolve("hg38_ensembl.ser");
+    }
+
     public Path transcriptCacheFor(GenomeBuild genomeBuild, TranscriptDatabase txDb) {
         return switch (genomeBuild) {
             case HG19 -> switch (txDb) {
                 case UCSC -> hg19UcscTxDatabase();
+                case ENSEMBL -> hg19EnsemblTxDatabase();
                 case REFSEQ -> hg19RefseqTxDatabase();
+                case REFSEQ_CURATED -> hg19RefseqCuratedTxDatabase();
             };
             case HG38 -> switch (txDb) {
                 case UCSC -> hg38UcscTxDatabase();
+                case ENSEMBL -> hg38EnsemblTxDatabase();
                 case REFSEQ -> hg38RefseqTxDatabase();
+                case REFSEQ_CURATED -> hg38RefseqCuratedTxDatabase();
             };
         };
     }